requirements.txt
fastapi
fastapi-sqlalchemy
psycopg2-binary
pydantic[dotenv]==1.9.1
uvicorn==0.18.2
alembic==1.8.1
jinja2==3.1.2
aiofiles
google-api-python-client==2.53.0
google-auth-httplib2
google-auth-oauthlib==0.5.2
SQLAlchemy==1.4.39
asyncio
pytz
icalendar==4.1.0
aiohttp
aiogoogle